agents/foo-bar/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "duckdb>=1.4.2",
    "duckduckgo-search>=8.1.1",
    "gitpython>=3.1.45",
    "langchain>=1.0.7",
    "langchain-anthropic>=1.0.4",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.3",
    "langchain-google-vertexai>=3.0.3",
    "langchain-graph-retriever>=0.8.0",
    "langchain-kuzu>=0.4.2",
    "langchain-mongodb>=0.7.1",
    "langchain-openai>=1.0.3",
    "langchain-postgres>=0.0.16",
    "langchain-qdrant>=1.1.0",
    "langgraph-cli[inmem]>=0.4.7",
    "meilisearch>=0.38.0",
    "openapi-schema-pydantic>=1.2.4",
    "outlines>=1.2.8",
    "promptlayer>=1.0.75",
    "redis>=6.4.0",
    "redisvl>=0.11.0",
    "supabase>=2.24.0",
    "typesense>=1.3.0",
]
